requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
coloredlogs==15.0.1
humanfriendly==10.0
idna==3.3
numpy==1.21.4
paho-mqtt==1.6.1
pandas==1.3.4
postgres==4.0
psycopg2-binary==2.9.2
psycopg2-pool==1.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
